shared/requirements.txt
# shared/ and db/ dependencies — used by both api and pipeline
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg[binary]>=3.0.0
qdrant-client>=1.9.0
aio-pika>=9.4.0
httpx>=0.27.0

# LLM provider SDKs
ollama>=0.4.0
groq>=0.5.0
google-genai>=1.0.0
openai>=1.0.0
